levelmeter error handling

+++ status.h
@@ -35,7 +35,7 @@ _Bool CHECK_ERROR(enum StatusBit Bit);
 void CLEAR_ALL(void);
 
 void set_status_byte(uint16_t status_byte);
-uint16_t get_status_byte(void);
+uint8_t get_status_byte(void);
